+++ packages/core/src/objects/Maybe.ts
@@ -1,4 +1,5 @@
 import type { PDFDictionary } from "./Dictionary";
+import { PDFNull } from "./Null";
 
 export class Maybe<T extends PDFObject> {
 
@@ -16,7 +17,7 @@ export class Maybe<T extends PDFObject> {
    * @returns returns internal value
    */
   public get(required = false, compressed?: boolean): T {
-    if (!this.parent.has(this.name)) {
+    if (!this.parent.has(this.name) || this.parent.get(this.name) instanceof PDFNull) {
       if (required) {
         throw new Error(`Cannot get required field '${this.name}'. Field is empty.`);
       }