+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.2](https://github.com/PeculiarVentures/pdf/compare/v1.5.1...v1.5.2) (2023-01-17)
+
+**Note:** Version bump only for package @peculiarventures/pdf-cli
+
+
+
+
+
+## [1.5.1](https://github.com/PeculiarVentures/pdf/compare/v1.5.0...v1.5.1) (2023-01-15)
+
+**Note:** Version bump only for package @peculiarventures/pdf-cli
+
+
+
+
+
+# [1.5.0](https://github.com/PeculiarVentures/pdf/compare/v1.4.0...v1.5.0) (2022-12-27)
+
+**Note:** Version bump only for package @peculiarventures/pdf-cli
+
+
+
+
+
 # [1.4.0](https://github.com/PeculiarVentures/pdf/compare/v1.3.0...v1.4.0) (2022-12-05)
 
 **Note:** Version bump only for package @peculiarventures/pdf-cli